added TableHead, TableBody as separated components
