+++ client/src/pages/Request/Request.tsx
@@ -1,29 +1,8 @@
-import {
-  Paper,
-  TableContainer,
-  Table,
-  TableBody,
-  TableHead,
-  TableRow,
-  TableCell,
-  Grid,
-  Typography,
-  Avatar,
-  Chip,
-} from '@material-ui/core';
+import { Paper, TableContainer, Table, Grid, Typography } from '@material-ui/core';
 import useStyles from './useStyles';
 
-import RequestModal from './RequestModal';
-
-function OwnerInfo(): JSX.Element {
-  const classes = useStyles();
-  return (
-    <Grid container alignItems="center">
-      <Avatar alt="avatar" src="" className={classes.userAvatar} />
-      <Typography className={classes.userDisplayName}>Joe Snow</Typography>
-    </Grid>
-  );
-}
+import RequestTableHead from './RequestTableHead';
+import RequestTableBody from './RequestTableBody';
 
 export default function Request(): JSX.Element {
   const classes = useStyles();
@@ -39,28 +18,8 @@ export default function Request(): JSX.Element {
         <Grid item>
           <TableContainer>
             <Table>
-              <TableHead>
-                <TableRow>
-                  <TableCell>Ower Info</TableCell>
-                  <TableCell align="center">Status</TableCell>
-                  <TableCell align="center">Price</TableCell>
-                  <TableCell align="center">Date created</TableCell>
-                  <TableCell align="center">Action</TableCell>
-                </TableRow>
-              </TableHead>
-              <TableBody>
-                <TableRow>
-                  <TableCell align="center">
-                    <OwnerInfo />
-                  </TableCell>
-                  <TableCell align="center">
-                    <Chip label="Pending" />
-                  </TableCell>
-                  <TableCell align="center">$35</TableCell>
-                  <TableCell align="center">2 days ago</TableCell>
-                  <RequestModal />
-                </TableRow>
-              </TableBody>
+              <RequestTableHead />
+              <RequestTableBody />
             </Table>
           </TableContainer>
         </Grid>
